soc_copilot_function/requirements.txt
google-cloud-storage
google-cloud-aiplatform
google-cloud-bigquery
vertexai
